Añadido validación DNI + cambios toString()
Se a desarrollado:

- Validación DNI que controla que un DNI sea válido y que no se encuentre dentro de la base de datos repetido.

- Se ha modificado el método toString() de las clases pojo con tal de que sea más legible y visual.

+++ src/main/java/view/ReservationManagerView.java
@@ -56,7 +56,19 @@ public class ReservationManagerView {
             int reservationId = Integer.parseInt(keyboard.nextLine());
 
             if (controller.doesReservationExist(reservationId)) {
-                isDeleted = controller.deleteReservationById(reservationId);
+
+                System.out.println("Detalles de la reserva a cancelar:");
+                System.out.println(controller.getReservationById(reservationId));
+
+                System.out.print("¿Está seguro de que desea cancelar esta reserva? (S/N): ");
+                String confirmation = keyboard.nextLine().trim().toUpperCase();
+
+                if (!confirmation.equals("S")) {
+                    System.out.println("Operación cancelada por el usuario.");
+                    isDeleted = false;
+                } else {
+                    isDeleted = controller.deleteReservationById(reservationId);
+                }
             } else {
                 System.out.println("Error! La reserva con ID " + reservationId + " no existe.");
             }
